When doing CHAP auth, prefer MS-CHAPv2 to ordinary CHAP.
The reason for a change is that this way one could truly negotiate
MPPE preferences (see "mppe-optional" patch) when UI does not offer
anything more verbose than "use encryption" checkbox (like iPad and
Android tablets).

If ordinary CHAP is preferred, clients succesfully authenticate,
then fail because of the encryption requirement. That's why the
preference change was required.

On the other hand, if server allows, but client does not support
(or just don't want to use) MS-CHAPv2, they'll just ConfNak it and
succesfully use CHAP.